put redux files in seperate folder, added is owner is true in nav
