+++ src/components/Navigation/index.js
@@ -16,13 +16,15 @@ export default function Navigation() {
   return (
     <Navbar bg="light" expand="lg">
       <Navbar.Brand as={NavLink} to="/">
-        YOUR PROJECT NAME
+        Be a Mayor
       </Navbar.Brand>
       <Navbar.Toggle aria-controls="basic-navbar-nav" />
       <Navbar.Collapse id="basic-navbar-nav">
         <Nav style={{ width: "100%" }} fill>
           <NavbarItem path="/" linkText="Home" />
-          {/* {token ? <NavbarItem path="/addacity" linkText="Add a city" /> : null} */}
+          {/* {token && isOwner === true ? (
+            <NavbarItem path="/addacity" linkText="Add a city" />
+          ) : null} */}
           <NavbarItem path="/addacity" linkText="city" />
           {loginLogoutControls}
           <span>
