+++ src/store/addACity/selectors.js
@@ -1,0 +1,1 @@
+export const selectAddCity = (state) => state.user;
