+++ src/store/user/reducer.js
@@ -1,9 +1,4 @@
-import {
-  LOG_OUT,
-  LOGIN_SUCCESS,
-  TOKEN_STILL_VALID,
-  ADD_CITY_ADDED,
-} from "./actions";
+import { LOG_OUT, LOGIN_SUCCESS, TOKEN_STILL_VALID } from "./actions";
 
 const initialState = {
   token: localStorage.getItem("token"),
@@ -22,9 +17,6 @@ export default (state = initialState, action) => {
       return { ...initialState, token: null };
 
     case TOKEN_STILL_VALID:
-      return { ...state, ...action.payload };
-
-    case ADD_CITY_ADDED:
       return { ...state, ...action.payload };
 
     default:
